Fix type casting in add_tensor_pre_hook
Currently triggers warning C4267 on MSVC with all warnings enabled. Add static_cast<int> to silence warning

+++ torch/csrc/dynamo/compiled_autograd.h
@@ -663,7 +663,7 @@ class CompiledNodeArgs {
   void add_tensor_pre_hook(c10::SafePyObject&& obj, int index) {
     auto fn_id = _compiler.emplace_hook(std::move(obj));
     collect_size(fn_id);
-    _node_call.tensor_pre_hooks.emplace_back(fn_id, index);
+    _node_call.tensor_pre_hooks.emplace_back(static_cast<int>(fn_id), static_cast<int>(index));
   }
 
   void add_cpp_single_tensor_pre_hook(
